Unify tests for compiler command in the same file
Tests for the "spack compiler" command were previously
scattered among different files.
